+++ src/main/java/learn/springang/umsportal/utils/JwtProvider.java
@@ -1,0 +1,88 @@
+package learn.springang.umsportal.utils;
+
+import com.auth0.jwt.JWT;
+import com.auth0.jwt.JWTVerifier;
+import com.auth0.jwt.algorithms.Algorithm;
+import com.auth0.jwt.exceptions.JWTVerificationException;
+import learn.springang.umsportal.domain.UserPrincipal;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.security.web.authentication.WebAuthenticationDetailsSource;
+import org.springframework.stereotype.Component;
+import org.springframework.util.StringUtils;
+
+import javax.servlet.http.HttpServletRequest;
+import java.util.Date;
+import java.util.List;
+import java.util.stream.Collectors;
+
+import static learn.springang.umsportal.constants.SecurityConstants.*;
+
+@Component
+public class JwtProvider {
+
+    @Value("{jwt.secret}")
+    private String secret;
+
+    public String generateJwtToken(UserPrincipal userPrincipal) {
+        String[] claims = getClaimsFromUser(userPrincipal);
+        return JWT.create()
+                .withIssuer(COMPANY_LLC)
+                .withAudience(COMPANY_ADMINISTRATION)
+                .withIssuedAt(new Date())
+                .withSubject(userPrincipal.getUsername())
+                .withArrayClaim(AUTHORITIES, claims)
+                .withExpiresAt(new Date(System.currentTimeMillis() + EXPIRATION_TIME))
+                .sign(Algorithm.HMAC512(secret));
+    }
+
+    public List<GrantedAuthority> getAuthorities(String token) {
+        List<String> claims = getClaimsFromToken(token);
+        return claims.stream().map(SimpleGrantedAuthority::new).collect(Collectors.toList());
+    }
+
+    public Authentication getAuthentication(String username, List<GrantedAuthority> authorities, HttpServletRequest request) {
+        UsernamePasswordAuthenticationToken authentication = new UsernamePasswordAuthenticationToken(
+                username, null, authorities);
+        authentication.setDetails(new WebAuthenticationDetailsSource().buildDetails(request));
+        return authentication;
+    }
+
+    public boolean isTokenValid(String username, String token) {
+        JWTVerifier verifier = getJWTVerifier();
+        return StringUtils.hasText(username) && isTokenNotExpired(verifier, token);
+    }
+
+    private String[] getClaimsFromUser(UserPrincipal userPrincipal) {
+        return userPrincipal.getAuthorities().stream()
+                .map(GrantedAuthority::getAuthority)
+                .toArray(String[]::new);
+    }
+
+    private List<String> getClaimsFromToken(String token) {
+        JWTVerifier verifier = getJWTVerifier();
+        return verifier.verify(token).getClaim(AUTHORITIES).asList(String.class);
+    }
+
+    private JWTVerifier getJWTVerifier() {
+        try {
+            Algorithm algorithm = Algorithm.HMAC512(secret);
+            return JWT.require(algorithm).withIssuer(COMPANY_LLC).build();
+        } catch (JWTVerificationException e) {
+            throw new JWTVerificationException(TOKEN_CANNOT_BE_VERIFIED);
+        }
+    }
+
+    private boolean isTokenNotExpired(JWTVerifier verifier, String token) {
+        Date expiration = verifier.verify(token).getExpiresAt();
+        return expiration.before(new Date());
+    }
+
+    public String getSubject(String token) {
+        JWTVerifier verifier = getJWTVerifier();
+        return verifier.verify(token).getSubject();
+    }
+}