+++ history.md
@@ -2,6 +2,10 @@ History
 -------
 In reversed order.
 
+* JwtProvider::isTokenValid.
+* JwtProvider::getAuthentication.
+* JwtProvider::getAuthorities.
+* JwtProvider::generateJwtToken.
 * SecurityConstants.
 * Add yaml properties for h2 and mysql.
 * Use mysql_db_users.sql script to create DB and user.